Updates #266 homepage tile iterations
I'm open to the suggestion of increasing the number of tiles to three,
to make them equally divisible by three as by four. I think twelve
tiles makes the layout more busy than I wanted it to be. For
comparison, I've attached two screen shots showing eight tiles vs
twelve tiles.

I can think of a couple of alternatives. One might be to switch the
tiles to a two column grid in a tablet view (see third attachment), or
another possibility for us to consider is, like Hawaii (or Windows 8),
we can have some tiles extend to the width of two tiles, for emphasis.

It may be best to table the topic until we finalize the links we want
to display on the home view, but I'm willing to change something now if
we want. I'd like to close this issue.