Remove index.php from URL handling
Even through we implemented clean URLs, routes going to index.php directly are still supported.  What's worse, if these routes are the first ones generated after a cache clear, then they stick in the cached output.

This rewrite rule rewrites all URLs with index.php, removing the index.php from the URL.  This had to be done with care, as we do want to still redirect to index.php.  So we're careful to only match client-requested routes.  This should leave the internal redirect routes alone.